Remove unnecessary session for webcal
Since there's no account authentication, we don't need a session
